feat: move NotFoundException and handler to common utils
