+++ utils/exception-handler/src/main/java/com/pda/exceptionhandler/GlobalExceptionHandler.java
@@ -2,10 +2,7 @@ package com.pda.exceptionhandler;
 
 import com.pda.apiutils.ApiUtils;
 import com.pda.apiutils.GlobalExceptionResponse;
-import com.pda.exceptionhandler.exceptions.BadRequestException;
-import com.pda.exceptionhandler.exceptions.ConflictException;
-import com.pda.exceptionhandler.exceptions.ForbiddenException;
-import com.pda.exceptionhandler.exceptions.UnAuthorizedException;
+import com.pda.exceptionhandler.exceptions.*;
 import io.swagger.v3.oas.annotations.Hidden;
 import org.springframework.context.support.DefaultMessageSourceResolvable;
 import org.springframework.http.HttpStatus;
@@ -50,6 +47,12 @@ public class GlobalExceptionHandler {
     @ExceptionHandler(ForbiddenException.class)
     @ResponseStatus(HttpStatus.FORBIDDEN)
     public GlobalExceptionResponse<String> handleForbiddenException(final ForbiddenException e) {
+        return ApiUtils.exception(e.getMessage());
+    }
+
+    @org.springframework.web.bind.annotation.ExceptionHandler(NotFoundException.class)
+    @ResponseStatus(HttpStatus.NOT_FOUND)
+    public GlobalExceptionResponse<String> handleNotFoundException(final NotFoundException e) {
         return ApiUtils.exception(e.getMessage());
     }
 }
